Move iterative rule tests from rule/test/ to rule/
The rule.test package is for the testing framework, not for the tests
themselves.
